Add automatic refresh of the session token
